Refactor acceleration to float and clean up debug output
Changed acceleration from int to float in sensor data structures and related functions for improved precision. Updated getMotionData to return acceleration in m/s² as float, adjusted gravity offset, and applied a small calibration to angular value. Removed verbose voltage debug output from Firebase transmission. Added empty wifi_serial and debug header files.

+++ ESP32-S3-Main/src/communication/firebase_manager.h
@@ -6,7 +6,7 @@
 
 struct FirebaseTxData {
     // Sensor readings - SEND ONLY
-    int acceleration;           // Send acceleration
+    float acceleration;           // Send acceleration
     int angular;               // Send angular velocity
     int battery;               // Send battery level
     float voltage;             // Send voltage
