+++ ESP32-S3-Main/src/sensors/mpu6050.h
@@ -54,7 +54,7 @@ public:
 
     int getAccelerationInt();
     int getAngularVelocityInt();
-    void getMotionData(int& acceleration, int& angular);
+    void getMotionData(float& acceleration, int& angular);
 };
 
 #endif
